add setupShaderEnv
* optionally injects code when '#pragma env' is found within shader
code. This is useful for 'uebershader' scenarios.

* This is a useful container for shader-specific `#define` statements.
If you want to re-use the same GLSL codebase for many permutations of a
shader, e.g. in an uebershader scenario, you may then activate or
de-activate code-paths by setting the appropriate preprocessor variables
with #defines using this method, before combining the shader.

* `#pragma env` is evaluated immediately after `#pragma include`, and
just before the shader is compiled by the graphics driver.