overhaul Context
+ add generic setters and getters for uniforms
+ add const& getters for uniforms
+ add dynamic CPU memory backing to uniforms, derived from shader
  inspection
+ automatically allocate correct CPU memory for
+ add stacks to uniform groups (allow for per-ubo push/pop)
+ only write to GPU when recording draw calls (lazy GPU memory
  allocation and write)
+ add memory structure which represents relations between
  descriptorsets, uniforms, and uniform members

Context is now better at mirroring current uniform state - it allows you
to get and set uniform state at any point without roundtrip to GPU, all
uniforms are cached within context.

Uniforms are also stacked within context, and UBO groups may be pushed
and popped from the CPU stack - this is mirrored on the GPU and no extra
memory is allocated if a popped matrix was already written to GPU
memory.

As uniform stacks are by binding, you can define how large a push/pop
group is by defining these in your shader - everything within an ubo can
be pushed and popped by using the UBO block name as identifier for

	context.pushBuffer(ubo_)

You can then query your popped uniforms by calling:

	context.getUniform<ofMatrix4x4>( "viewMatrix");

Notice how this is a typed method - if the type for the uniform doesn't
match the memory allocated through the shader you will receive an error
log - as you will if an uniform with that name has not been found.