fix an embarrassing mistake
Allocator would not return the correct GPU buffer offset, but would
point GPU memory all onto offset 0, effectively the first frame. This
lead to some ugly crashes, if frame 0 and 1 differ in memory footprint.
It also lead to the impression of vsync problems as the frame was
effectively drawn while its memory was being written to.