add dynamic generation of pipelines - and pipeline caching
+ if a pipeline state change is detected because of a dirty pipeline
flag, first the pipeline hash is calculated for the current pipeline
state. if a pipeline with such hash is not found in the cache, a new
pipeline is created, based on the currently bound pipeline. the new
pipeline is stored in the chache.

+ implements ofFill() ofNoFill()
  This is an example of two different pipelines needed for subtle
  difference.

Allocate descriptorSets on demand -

+ when we draw - and we have never drawn with this descriptorSetLayout,
  allocate the Descriptors necessary.

+ Keep one DescriptorPool per swapchain image - we're effectively
  multi-buffering descriptors. This allows us to just reset all
  descriptors per frame - and not to get into DescriptorPool
  fragmentation issues.

+ TODO: we should make space in the pool for a few more descriptors,
especially, if there are image decriptors. Each image descriptor needs
to have its buffer info baked in, so we can't just recycle these as
dynamic descriptors as we do for our default UBO descriptors. On the
other side, descriptors don't take away a lot of space (som sources say
32 bytes per) and we could make space for 1000s in one pool if we need
to.