add shaderc
+ this allows to process & compile glsl shaders into spir-v, which is
then further processed through spirv-cross.

+ glsl compilation is optional, and spirv input directly still works

+ by using shaderc we can make use of shader #include directives, and
other #pragmas that shaderc supports.

+ to use shaderc, you need to compile shaderc via apothecary - the
recipe for vs/64 works, but it appears that it's tricky to use the
release .lib for both Debug and Release targets - somehow, this seems to
work with the glfw .lib, which may be because glfw compiles as pure C
and not C++ as shaderc does...