rewrite api to be more idiomatic
+ remove context, remove implicit renderer methods

all rendering needs to be explicit and routed through the correct
contexts, also, draw commands hold their pipeline state, and only have
mutable elements pertaining to uniform and attribute input.

This way, we prepare for a batching renderer, which will be the next
developmental step down the line.

Also, the current organisation in context scopes is far from ideal -
this needs to change, and renderpass and subpass must become part of the
drawcommand declaration. This makes it possible to issue draw commands
in a more flexible way.