refactor: consolidate renderContext
A rendercontext has exactly one renderpass, and owns per-virtual frame
resources. it is the sole owner of a renderpass, and owns per-frame
resources exclusively.

A renderbatch can borrow a vkCommandBuffer from the renderpass, but
such a commandbuffer must be returned within the current frame.

Currently, only primary commandbuffers are allowed, but it's worth
thinking about being able to hand out secondary command buffers. Note
that a primary commandBuffer always begins with beginning a Renderpass.