refactor: renderpass into renderbatch
+ as a `RenderBatch` begins and ends a `vkRenderPass`, it makes more
sense to keep the renderpass with the `RenderBatch`, and remove any
references to renderpasses from the `Context`.

This means, a context may be used to create multiple `renderBatch`es,
and each `RenderBatch` may render to its own framebuffer/use its own
renderpass.

Remember, in the end a `RenderBatch` is reduced to a `vkCommandBuffer`.