refactor: make allocator standalone objects
Allocators can now be stored as members without unique_ptr wrapper in
other objects.

Allocators follow the pattern of classes which need `setup()` to be
fully functional. Every class which has a `Settings` struct should only
be used after it has been setup using the settings struct. Requiring
such classes to be wrapped in unique ptrs is overkill, and only
confuses.

The recommended pattern for Allocators is now to store them as regular
members in other classes, and to call setup on them in the host class
when the host class is setup. To enforce destruction order, you may call
`.reset` on the Allocator on host class teardown. Reset is also called
in the Allocator's destructor.