vk::Texture refactor
+ refactor to 'deferred initialisation' pattern.

This adds a Settings object to Texture, which will initialise to sane
defaults. Settings can then be modified, and Texture can be
allocated/setup using `Texture::setup(settings)`.

I'm not 100% happy with this pattern, as it does not allow us to
auto-destroy internal vk objects, but it makes parameters more
discoverable, and reduces verbosity in application code.